+++ internal/server/log.go
@@ -14,20 +14,20 @@ func NewLog() *Log {
 	return &Log{}
 }
 
-func (c *Log) Append(record Record) (int, error) {
+func (c *Log) Append(record Record) (uint64, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	record.Offset = len(c.records)
+	record.Offset = uint64(len(c.records))
 	c.records = append(c.records, record)
 	return record.Offset, nil
 }
 
-func (c *Log) Read(offset int) (Record, error) {
+func (c *Log) Read(offset uint64) (Record, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	if offset >= len(c.records) {
+	if offset >= uint64(len(c.records)) {
 		return Record{}, ErrOffsetNotFound
 	}
 	return c.records[offset], nil
@@ -35,7 +35,7 @@ func (c *Log) Read(offset int) (Record, error) {
 
 type Record struct {
 	Value  []byte `json:"value"`
-	Offset int    `json:"offset"`
+	Offset uint64 `json:"offset"`
 }
 
 var ErrOffsetNotFound = fmt.Errorf("offset not found")
